Fix #33
The fault was Node the entire time. I thought the problem was with writing the modified file to the disk so I spent all this time messing around with different formats and whatnot. It turns out if readFileSync doesn't understand the encoding it auto converts whatever it can to utf-8. But since it doesn't understand the other encodings this breaks the characters. My solution was a package called `iconv-lite` that has support for these other formats and can handle them properly.